refactor: 코드 재정렬 및 디렉토리 구조 수정
