+++ src/main/java/teamyc/recordpet/global/security/SecurityConfig.java
@@ -1,4 +1,4 @@
-package teamyc.recordpet.global.config;
+package teamyc.recordpet.global.security;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -7,6 +7,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 
 @Configuration
 public class SecurityConfig {
+
     @Bean
     public PasswordEncoder passwordEncoder() {
         return new BCryptPasswordEncoder();
